Minor tweaks picked up in code review.
