+++ src/briefcase/commands/new.py
@@ -12,10 +12,10 @@ from briefcase.config import (
     is_valid_bundle_identifier,
     make_class_name,
 )
-from briefcase.exceptions import TemplateUnsupportedVersion
+from briefcase.exceptions import BriefcaseCommandError, TemplateUnsupportedVersion
 from briefcase.integrations import git
 
-from .base import BaseCommand, BriefcaseCommandError
+from .base import BaseCommand
 
 
 def titlecase(s):
